requirements.txt
# Data Ingestion Requirements

# Core data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Financial data
yfinance>=0.2.28
pandas-datareader>=0.10.0
fredapi>=0.5.0  # Fallback for FRED data (not currently used)
requests>=2.28.0  # For direct FRED API calls

# Configuration
pyyaml>=6.0

# Logging
loguru>=0.7.0

# HDF5 support
tables>=3.8.0

# Type checking (optional but recommended)
typing-extensions>=4.7.0

# Machine Learning
lightgbm>=4.0.0
interpret>=0.4.0
imodels>=1.3.0
shap>=0.44.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0  # For Hidden Markov Models (regime detection)
